requirements.txt
Flask
requests
python-telegram-bot
boto3
